chore: #251 뷰 이사가기
