+++ Spoony-iOS/Spoony-iOS/Source/Feature/MyPage/Attendance/StampCategoryRow.swift
