+++ Spoony-iOS/Spoony-iOS/Source/Feature/MyPage/MyPageView.swift
@@ -51,7 +51,7 @@ struct MyPageView: View {
             case let .settings(store):
                 SettingsView(store: store)
             case let .attendance(store):
-                MealTrackerView(store: store)
+                AttendanceView(store: store)
                     .navigationBarBackButtonHidden()
                     .toolbar(.hidden, for: .tabBar)
